+++ packages/material-composer-r3f/src/reactor.tsx
@@ -1,8 +1,9 @@
 import { Module, ModuleFactory, ModuleFactoryProps } from "material-composer"
 import * as Modules from "material-composer/modules"
-import { FC, useMemo, useRef } from "react"
+import { FC, useMemo } from "react"
 import { Input } from "shader-composer"
 import { useModuleRegistration } from "./moduleRegistration"
+import { useDetectShallowChange } from "./lib/useDetectShallowChange"
 
 type Modules = typeof Modules
 
@@ -20,28 +21,12 @@ type ModuleComponentProxy = {
     : never
 }
 
-const usePropsHaveChanged = (props: any) => {
-  const version = useRef(0)
-  const previousProps = useRef<any>()
-
-  const changed =
-    !previousProps.current ||
-    Object.keys(props).length !== Object.keys(previousProps.current).length ||
-    Object.keys(props).some((key) => {
-      return !Object.is(previousProps.current[key], props[key])
-    })
-
-  previousProps.current = props
-
-  return changed ? ++version.current : version
-}
-
 export const makeModuleComponent = <P extends ModuleFactoryProps>(
   fac: ModuleFactory<P>
 ) => (props: P) => {
   const module = useMemo(() => {
     return fac(props)
-  }, [usePropsHaveChanged(props)])
+  }, [useDetectShallowChange(props)])
 
   useModuleRegistration(module)
